+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.5] - 2018-11-30
 ### Added
 - `SQLiteCursor#getDateTime()` method
-- `SQLiteDatatypeMismatchException` exception
+- `SQLiteDatabase#validateSql()` method
 - `SQLiteDatabase#yieldIfContendedSafely()` method
+- `SQLiteDatatypeMismatchException` exception
 
 ## [0.1.4] - 2018-11-30
 ### Added